refactor: split up releaseData into separate files
