+++ src/assets/data/betaReleases.ts
@@ -1,0 +1,98 @@
+export type ReleaseInfo = {
+  name: string;
+  browser_download_url: string;
+  checksum: string;
+  type: string;
+};
+
+type ReleaseDirectory = {
+  version: string;
+  win: ReleaseInfo[];
+  mac: ReleaseInfo[];
+  lin: ReleaseInfo[];
+  src: ReleaseInfo[];
+};
+
+export const betaReleases: ReleaseDirectory = {
+  version: "3.5.0 beta",
+  win: [
+    {
+      name: "64 bit installer (recommended)",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-win-3.5.0-beta-3-64bit.exe",
+      checksum:
+        "ed7de964ed11cbc8f74e815dbcb2cb8487ba136818cf1a148f16cadd4c10f3d0",
+      type: ".exe",
+    },
+    {
+      name: "64 bit zip file",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-win-3.5.0-beta-3-64bit.zip",
+      checksum:
+        "52da5c3e2507408d72c4ab425c2e465e3c8ad452b2ac89ddfb3f5bc141d68a03",
+      type: ".zip",
+    },
+    {
+      name: "32 bit installer",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-win-3.5.0-beta-3-32bit.exe",
+      checksum:
+        "6bb6c0d3513be7d98c400f43d84cd39992065f4c6460d80b6cb1667733ca95c7",
+      type: ".exe",
+    },
+    {
+      name: "32 bit zip file",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-win-3.5.0-beta-3-32bit.zip",
+      checksum:
+        "c313ca3c475b487bf88a42537cbc9454090391250017fe210226b3ca78797d9a",
+      type: ".zip",
+    },
+  ],
+  mac: [
+    {
+      name: "Universal dmg (recommended)",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-macOS-3.5.0-beta-3-universal.dmg",
+      checksum:
+        "9500ede91b837fc12e5106fa33d6603829288b90fb1e28d2d70bfee9db33406e",
+      type: ".dmg",
+    },
+    {
+      name: "ARM 64 dmg (Apple Silicon)",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-macOS-3.5.0-beta-3-arm64.dmg",
+      checksum:
+        "0f3e9b9ee8e77d8b8613db8d66927e982bbec870e801811060d3d8fbc25c7698",
+      type: ".dmg",
+    },
+    {
+      name: "x86_64 dmg (Intel)",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-macOS-3.5.0-beta-3-x86_64.dmg",
+      checksum:
+        "415342de27b572bd3801f51bd77e850a21701b39e2392c2c347ea8db4da4f122",
+      type: ".dmg",
+    },
+  ],
+  lin: [
+    {
+      name: "AppImage",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-linux-3.5.0-beta-3-x64.AppImage",
+      checksum:
+        "fc5df63e3819f4f59b4366c436579bd1a73b045a4dae28316edf6c23948a06ce",
+      type: ".AppImage",
+    },
+  ],
+  src: [
+    {
+      name: "Source code",
+      browser_download_url:
+        "https://github.com/audacity/audacity/releases/download/Audacity-3.5.0-beta-3/audacity-sources-3.5.0-beta-3.tar.gz",
+      checksum:
+        "53c33ed875f4eb1501707d7989a6de253370a368c1c50a877e5cfa96c02bebdc",
+      type: ".tar.gz",
+    },
+  ],
+};
