+++ src/pages/beta.astro
@@ -1,7 +1,7 @@
 ---
 import BaseLayout from "../layouts/BaseLayout.astro";
 import "../styles/icons.css";
-import { betaReleases } from "../assets/js/releaseData";
+import { betaReleases } from "../assets/data/betaReleases";
 import SplitDownloadButton from "../components/button/SplitDownloadButton.jsx";
 
 const { win, mac, lin } = betaReleases;
