+++ src/pages/download.astro
@@ -2,7 +2,7 @@
 import BaseLayout from "../layouts/BaseLayout.astro";
 import OperatingSystemCard from "../components/card/OperatingSystemCard";
 import "../styles/icons.css";
-import { audacityReleases } from "../assets/js/releaseData";
+import { audacityReleases } from "../assets/data/audacityReleases";
 import ChecksumAccordion from "../components/accordion/ChecksumAccordion";
 
 const { version, src } = audacityReleases;
