+++ src/pages/download/linux.astro
@@ -1,6 +1,6 @@
 ---
 import DownloadPageLayout from "../../layouts/DownloadPageLayout.astro";
-import { audacityReleases } from "../../assets/js/releaseData";
+import { audacityReleases } from "../../assets/data/audacityReleases";
 
 const { lin, version } = audacityReleases;
 ---
