+++ src/pages/download/mac.astro
@@ -1,6 +1,7 @@
 ---
 import DownloadPageLayout from "../../layouts/DownloadPageLayout.astro";
-import { audacityReleases, museHubReleases } from "../../assets/js/releaseData";
+import { audacityReleases } from "../../assets/data/audacityReleases";
+import { museHubReleases } from "../../assets/data/museHubReleases";
 
 const { mac, version } = audacityReleases;
 ---
