+++ src/pages/download/openvino.astro
@@ -1,8 +1,6 @@
 ---
-import {
-  openvinoReleases,
-  audacityReleases,
-} from "../../assets/js/releaseData";
+import { openvinoReleases } from "../../assets/data/openvinoReleases";
+import { audacityReleases } from "../../assets/data/audacityReleases";
 import "../../styles/icons.css";
 import BaseLayout from "../../layouts/BaseLayout.astro";
