+++ src/pages/download/windows.astro
@@ -1,5 +1,6 @@
 ---
-import { audacityReleases, museHubReleases } from "../../assets/js/releaseData";
+import { audacityReleases } from "../../assets/data/audacityReleases";
+import { museHubReleases } from "../../assets/data/museHubReleases";
 import "../../styles/icons.css";
 import DownloadPageLayout from "../../layouts/DownloadPageLayout.astro";
 ---
